Convert to generic compound and add uniswap.
