+++ client/scripts/app.ts
@@ -14,7 +14,7 @@ import moment from 'moment';
 import mixpanel from 'mixpanel-browser';
 import _ from 'underscore';
 
-import { AaveTypes, MarlinTypes, MolochTypes } from '@commonwealth/chain-events';
+import { AaveTypes, CompoundTypes, MolochTypes } from '@commonwealth/chain-events';
 import app, { ApiStatus, LoginState } from 'state';
 import {
   ChainInfo,
@@ -274,13 +274,13 @@ export async function selectNode(n?: NodeInfo, deferred = false): Promise<boolea
       './controllers/chain/ethereum/moloch/adapter'
     )).default;
     newChain = new Moloch(n, app);
-  } else if (MarlinTypes.EventChains.find((c) => c === n.chain.network)) {
-    const Marlin = (await import(
+  } else if (CompoundTypes.EventChains.find((c) => c === n.chain.network || c === n.chain.id)) {
+    const Compound = (await import(
       /* webpackMode: "lazy" */
-      /* webpackChunkName: "marlin-main" */
-      './controllers/chain/ethereum/marlin/adapter'
+      /* webpackChunkName: "compound-main" */
+      './controllers/chain/ethereum/compound/adapter'
     )).default;
-    newChain = new Marlin(n, app);
+    newChain = new Compound(n, app);
   } else if (AaveTypes.EventChains.find((c) => c === n.chain.network)) {
     const Aave = (await import(
       /* webpackMode: "lazy" */
