+++ client/scripts/models/types.ts
@@ -33,8 +33,7 @@ export enum ChainNetwork {
   Ethereum = 'ethereum',
   NEAR = 'near',
   Moloch = 'moloch',
-  Marlin = 'marlin',
-  MarlinTestnet = 'marlin-testnet',
+  Compound = 'compound',
   Aave = 'aave',
   AaveLocal = 'aave-local',
   dYdX = 'dydx',
@@ -50,7 +49,6 @@ export enum ChainNetwork {
   Crust = 'crust',
   CosmosHub = 'cosmoshub',
   Gaia13k = 'gaia-13k',
-  Dydx = 'dydx',
 }
 
 // This function returns a default chain for a chainbase
@@ -115,7 +113,7 @@ export enum VotingType {
   MultiOptionVoting = 'multioption',
   None = 'none',
   MolochYesNo = 'moloch',
-  MarlinYesNo = 'marlin',
+  CompoundYesNo = 'compound',
 }
 export enum VotingUnit {
   OnePersonOneVote = '1p1v',
