+++ server/scripts/migrateChainEntities.ts
@@ -11,8 +11,8 @@ import {
   chainSupportedBy,
   isSupportedChain,
   IStorageFetcher,
-  MarlinTypes,
-  MarlinEvents,
+  CompoundTypes,
+  CompoundEvents,
   MolochTypes,
   AaveTypes,
   AaveEvents,
@@ -63,9 +63,9 @@ export async function migrateChainEntity(chain: string): Promise<void> {
       // TODO: determine moloch API version
       // TODO: construct dater
       throw new Error('Moloch migration not yet implemented.');
-    } else if (chainSupportedBy(chain, MarlinTypes.EventChains)) {
-      const api = await MarlinEvents.createApi(node.url, node.address);
-      fetcher = new MarlinEvents.StorageFetcher(api);
+    } else if (chainSupportedBy(chain, CompoundTypes.EventChains)) {
+      const api = await CompoundEvents.createApi(node.url, node.address);
+      fetcher = new CompoundEvents.StorageFetcher(api);
       // range.startBlock = chain === 'aave' ? 12200000 : 0;
       range.startBlock = 0;
     } else if (chainSupportedBy(chain, AaveTypes.EventChains)) {
