+++ server/util/proposalIdToEntity.ts
@@ -1,4 +1,4 @@
-import { SubstrateTypes, MolochTypes, MarlinTypes, AaveTypes } from '@commonwealth/chain-events';
+import { SubstrateTypes, MolochTypes, CompoundTypes, AaveTypes } from '@commonwealth/chain-events';
 
 // this function takes an "old style" identifier such as treasuryproposal_4 and attempts
 // fetch the corresponding chain entity from the database
@@ -25,8 +25,8 @@ export default async function (models, chain: string, identifier: string) {
     case 'molochproposal': {
       return findEntity(MolochTypes.EntityKind.Proposal.toString());
     }
-    case 'marlinproposal': {
-      return findEntity(MarlinTypes.EntityKind.Proposal.toString());
+    case 'compoundproposal': {
+      return findEntity(CompoundTypes.EntityKind.Proposal.toString());
     }
     case 'aaveproposal': {
       return findEntity(AaveTypes.EntityKind.Proposal.toString());
