+++ shared/adapters/chain/compound/types.ts
@@ -1,0 +1,9 @@
+import { CompoundTypes } from '@commonwealth/chain-events';
+import { ICompletable } from '../../shared';
+
+export type ICompoundProposalResponse = Omit<CompoundTypes.IProposalCreated, 'kind'> & ICompletable & {
+  eta?: number;
+  queued: boolean;
+  executed: boolean;
+  cancelled: boolean;
+};
