+++ shared/eth/artifacts/contracts/Marlin/GovernorAlpha.sol/GovernorAlpha.json
@@ -1,7 +1,7 @@
 {
   "_format": "hh-sol-artifact-1",
   "contractName": "GovernorAlpha",
-  "sourceName": "contracts/Marlin/GovernorAlpha.sol",
+  "sourceName": "contracts/compound/GovernorAlpha.sol",
   "abi": [
     {
       "inputs": [
