+++ shared/eth/artifacts/contracts/Marlin/GovernorAlpha.sol/TimelockInterface.json
@@ -1,7 +1,7 @@
 {
   "_format": "hh-sol-artifact-1",
   "contractName": "TimelockInterface",
-  "sourceName": "contracts/Marlin/GovernorAlpha.sol",
+  "sourceName": "contracts/compound/GovernorAlpha.sol",
   "abi": [
     {
       "inputs": [],
